ロジックの切り出し
